Extended the RunChangeIncident to include the run number (avoid unwanted dependencies on ODIN).
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@5283 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 0c2096d5ed4fd172f26f32366436208473a6292c

+++ LHCb/Det/DetDescSvc/src/RunChangeHandlerSvc.h
@@ -1,4 +1,4 @@
-// $Id: RunChangeHandlerSvc.h,v 1.1 2009-02-03 18:31:02 marcocle Exp $
+// $Id: RunChangeHandlerSvc.h,v 1.2 2009-02-04 18:02:28 marcocle Exp $
 #ifndef _RUNCHANGEHANDLERSVC_H_
 #define _RUNCHANGEHANDLERSVC_H_
 
@@ -119,7 +119,7 @@ private:
   Conditions m_conditions;
   
   /// Current run number.
-  unsigned int m_currentRun;
+  unsigned long m_currentRun;
   
   /// EventDataSvc, for ODIN
   mutable IDataProviderSvc *m_evtSvc;
